[refact] : AddButton interface 분리
